DRUP-Itp: DRUP proof logging and traversal
Support for online DRUP proofs and proof traversal.
Online DRUP proofs are maintained in main memory (as opposed to
on disk) and can be effectively traversed, trimmed, and manipulated.

This enables efficient interpolation computation.

The proofs are not intended to be certifying, hence existing Glucose
BCP is used for validation / reconstruction / traversal.

Provides API for an external proof traversal to enable implementing
interpolation. See AVY model checker for details on interpolation
algorithm.

See DRUPing for Interpolants by Vizel and Gurfinkel for details of the
algorithm.